+++ composeApp/src/commonMain/kotlin/org/darthacheron/fitbe/settings/converters/FluidUnitConverter.kt
@@ -3,7 +3,7 @@ package org.darthacheron.fitbe.settings.converters
 import org.darthacheron.fitbe.health.beverages.FluidUnit
 
 class FluidUnitConverter {
-    fun toMilliliter(amount: UInt, from: FluidUnit): UInt {
+    fun toMilliliter(amount: Double, from: FluidUnit): Double {
         return from.toMilliliter(amount)
     }
 }
